Fixing media queries, not for notch devices
